Move shared packages out of the root
