+++ lib/manifest/collection.go
