+++ lib/manifest/http.go
