+++ lib/manifest/hypermode.json
