+++ lib/manifest/manifest.go
