+++ lib/manifest/model.go
