+++ lib/manifest/test/valid_hypermode.json
