+++ lib/wasmextractor/go.mod
@@ -1,0 +1,3 @@
+module github.com/hypermodeinc/modus/lib/wasmextractor
+
+go 1.23.0
