+++ lib/wasmextractor/wasmextractor.go
