+++ runtime/sqlclient/pooling.go
@@ -13,7 +13,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/hypermodeinc/modus/pkg/manifest"
+	"github.com/hypermodeinc/modus/lib/manifest"
 	"github.com/hypermodeinc/modus/runtime/manifestdata"
 	"github.com/hypermodeinc/modus/runtime/secrets"
